requirements.txt
absl-py==0.9.0
affine==2.3.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.12
appdirs==1.4.3
applaunchservices==0.2.1
appnope==0.1.0
argh==0.26.2
asn1crypto==1.3.0
astor==0.8.1
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.4
aws-requests-auth==0.4.2
Babel==2.8.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.2
bleach==3.1.0
boto3==1.12.2
botocore==1.15.2
branca==0.4.0
bs4==0.0.1
Cartopy==0.17.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
click-plugins==1.1.1
cligj==0.5.0
cloudpickle==1.3.0
clyent==1.2.2
colorlover==0.3.0
conda==4.8.3
conda-build==3.18.11
conda-package-handling==1.6.0
conda-verify==3.4.2
contextily==1.0rc2
cryptography==2.8
cufflinks==0.17.3
cycler==0.10.0
dash==1.9.0
dash-core-components==1.8.0
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.0
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.7
descartes==1.1.0
diff-match-patch==20181111
dnspython==1.16.0
docutils==0.16
entrypoints==0.3
esda==2.2.1
esy-osm-pbf==0.0.1
filelock==3.0.12
Fiona==1.8.13.post1
flake8==3.7.9
Flask-Compress==1.4.0
Flask-Cors==3.0.8
Flask-JSON==0.3.4
folium==0.10.1
future==0.18.2
gast==0.3.3
geographiclib==1.50
geopandas==0.7.0
geoplot==0.4.0
geopy==1.21.0
giddy==2.3.0
glob2==0.7
gmplot==1.2.0
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
haversine==2.2.0
idna==2.9
imagesize==1.2.0
importlib-metadata==1.5.0
inequality==1.0.0
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.15.2
Jinja2==2.11.1
jmespath==0.9.4
joblib==0.14.1
json5==0.9.3
jsonlines==1.2.0
jsonschema==3.2.0
jupyter-client==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.7
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==21.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
libarchive-c==2.8
libpysal==4.2.2
lief==0.9.0
mapclassify==2.2.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.0
mccabe==0.6.1
mercantile==1.1.2
mgwr==2.1.1
mistune==0.8.4
mongoengine==0.10.7
more-itertools==8.2.0
munch==2.5.0
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
notebook==6.0.3
numpy==1.18.2
numpydoc==0.9.2
olefile==0.46
osmnx==0.11.4
packaging==20.3
pandas==1.0.2
pandocfilters==1.4.2
parso==0.5.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
plotly==4.5.0
pluggy==0.13.1
pointpats==2.1.0
pomegranate==0.12.2
pprint==0.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
psutil==5.7.0
ptyprocess==0.6.0
py==1.8.1
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.20
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.6.1
pylint==2.4.4
pymongo==3.10.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyproj==2.4.2.post1
pyrsistent==0.15.7
pysal==2.2.0
pyshp==2.1.0
PySocks==1.7.1
pytest==5.4.1
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.9
pytz==2019.3
PyYAML==5.3.1
pyzmq==18.1.1
QDarkStyle==2.8
QtAwesome==0.7.0
qtconsole==4.7.2
QtPy==1.9.0
quantecon==0.4.6
quilt3==3.1.10
rasterio==1.1.2
rasterstats==0.14.0
requests==2.23.0
requests-futures==1.0.0
retrying==1.3.3
rope==0.16.0
Rtree==0.9.3
ruamel-yaml==0.15.87
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
scipy==1.4.1
seaborn==0.10.0
segregation==1.2.0
Send2Trash==1.5.0
Shapely==1.7.0
simplejson==3.17.0
six==1.14.0
snowballstemmer==2.0.0
snuggs==1.4.7
sortedcontainers==2.1.0
soupsieve==2.0
spaghetti==1.4.1
spglm==1.0.7
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spint==1.0.6
splot==1.1.2
spreg==1.0.4
spvcm==0.3.0
spyder==4.1.1
spyder-kernels==1.9.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tobler==0.2.0
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
ujson==1.35
urllib3==1.25.8
watchdog==0.10.2
watermark==2.0.2
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wurlitzer==2.0.0
xmltodict==0.12.0
yapf==0.28.0
zipp==2.2.0
